Created auto_correct class file.
Made the following changes to Dictionary:

Removed print_words_starting_with_letter as it's function was redundant and part of the other function.
Changed lookup_words name to is_word to better identify its functionality.
Added lookup_words function to return a list of found words.

Made the following changes to Trie:

Changed the original tree traversal's name to tree_traversal_with_print to better identify it.
Added a copy of tree_traversal and had it return a list of found words.