Created functionality to prompt for input file to correct and output file name to output corrected file.
Created functionality in main to search for possible corrections and print them to the user. Then take in the users selected correction and store that in a previously seen corrections list.

Created functionality to output corrected file and previously seen corrections.

All thats left to finish this pa is to add comments, documentation comments, and do a pa5 essay.